requirements.txt
uagents==0.23.4
aiohttp>=3.9.0
google-generativeai>=0.3.0
supabase>=2.0.0
python-dotenv>=1.0.0
croniter>=2.0.0
pydantic>=2.0.0
uagents-core==0.4.0
PyJWT>=2.8.0
bcrypt>=4.0.0
requests>=2.31.0
cryptography>=41.0.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-core>=0.1.0
langchain-community>=0.0.20
slack-sdk>=3.27.0
slack-bolt>=1.18.0
